# docs/api/openapi-v2.json
{"openapi":"3.0.1","info":{"title":"ImobLink API","description":"Documentação da API backend do projeto ImobLink","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"BearerAuth":[]}],"tags":[{"name":"Microsserviços","description":"Integração do backend com microsserviços de legenda e leitura de documento"},{"name":"Comments","description":"Realizar comentários em posts e perfis"},{"name":"Images","description":"Operações relacionadas a carregamento, save e busca de imagens"},{"name":"Messages","description":"Enviar mensgens e listar contatos"},{"name":"Feed","description":"Carregar posts"},{"name":"User","description":"Operações relacionadas ao perfil do usuário"},{"name":"Admin","description":"Operações que só podem ser realizadas por ADMIN's e SUPERADMIN"},{"name":"Authentication","description":"Operações de login e cadastro"},{"name":"Posts","description":"Operações relacionadas a posts"},{"name":"Follow","description":"Seguir perfis e listar perfis seguidos"}],"paths":{"/integracao/ocr":{"post":{"tags":["Microsserviços"],"summary":"Extair informações do RG","description":"Realiza a leitura do documento e preenche os campos de nome e CPF no cadastro","operationId":"processarDocumento","requestBody":{"content":{"application/json":{"schema":{"required":["frente","verso"],"type":"object","properties":{"frente":{"type":"string","format":"binary"},"verso":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/integracao/legenda":{"post":{"tags":["Microsserviços"],"summary":"Gerar legenda","description":"Gera legenda automática com base na imagem fornecida para o post","operationId":"gerarLegenda","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/posts/like/{id}":{"post":{"tags":["Posts"],"summary":"Curtir post","description":"Adiciona seu like ao post","operationId":"likePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/fav/{id}":{"post":{"tags":["Posts"],"summary":"Favoritar post","description":"Adiciona o post à sua lista de favoritos","operationId":"favPost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/create":{"post":{"tags":["Posts"],"summary":"Criar novo post","description":"Permite criar uma nova publicação com múltiplas imagens, descrição, preço, etc","operationId":"createPost","parameters":[{"name":"description","in":"query","required":true,"schema":{"type":"string"}},{"name":"price","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"street","in":"query","required":true,"schema":{"type":"string"}},{"name":"avenue","in":"query","required":true,"schema":{"type":"string"}},{"name":"number","in":"query","required":true,"schema":{"type":"string"}},{"name":"type","in":"query","required":true,"schema":{"type":"string"}},{"name":"images","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","format":"binary"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/addImage/{id}":{"post":{"tags":["Posts"],"summary":"Adicionar imagem","description":"Permite adicionar uma imagem à sua publicação","operationId":"addImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["image"],"type":"object","properties":{"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/messages/send/{id}":{"post":{"tags":["Messages"],"summary":"Enviar mensagem","description":"Permite enviar uma mensagem a outro usuário","operationId":"sendMessage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"content","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}},"security":[{"BearerAuth":[]}]}},"/api/images":{"post":{"tags":["Images"],"summary":"Salvar imagem","description":"Permite salvar endereços de imagens diretamente na db (mais utilizado para testes)","operationId":"saveImage","requestBody":{"content":{"application/json":{"schema":{"required":["image"],"type":"object","properties":{"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/follow/{followingId}":{"post":{"tags":["Follow"],"summary":"Seguir usuário","description":"Permite seguir o perfil de outro usuário","operationId":"followUser","parameters":[{"name":"followingId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/comments/comment/{userId}":{"post":{"tags":["Comments"],"summary":"Comentar em perfil de um user","description":"Permite realizar comentários no perfil de um usuário","operationId":"addComment","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentResponse"}}}}},"security":[{"BearerAuth":[]}]}},"/api/comments/comment/post/{postId}":{"post":{"tags":["Comments"],"summary":"Comentar em um post","description":"Permite realizar comentários em um post","operationId":"addCommentPost","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentResponse"}}}}},"security":[{"BearerAuth":[]}]}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Registrar nova conta","description":"Permite que o user crie uma conta na aplicação","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Acessar sua conta","description":"Permite que o user faça login na aplicação","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/adm/login":{"post":{"tags":["Authentication"],"summary":"Acessar conta de ADMIN na dashboard","description":"Permite que o ADMIN realize login na interface administrativa","operationId":"admLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/admin/promote":{"post":{"tags":["Admin"],"summary":"Promover USER para ADMIN","description":"Permite que o SUPERADMIN ou um ADMIN promova um usuário comum a administrador da plataforma","operationId":"promoteUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromoteRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/setPassword":{"patch":{"tags":["User"],"summary":"Alterar senha","description":"Permite alterar a senha da sua conta","operationId":"setPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/setInfo":{"patch":{"tags":["User"],"summary":"Editar informações da conta","description":"Permite editar as informações do seu perfil","operationId":"setInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetInfoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/setImageProfile":{"patch":{"tags":["User"],"summary":"Editar foto de perfil","description":"Permite alterar a sua foto de perfil","operationId":"setProfileImage","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["image"],"type":"object","properties":{"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/edit/{id}":{"patch":{"tags":["Posts"],"summary":"Editar post","description":"Permite editar as informações da sua publicação","operationId":"editPost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetPostInfoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/messages/edit/{mId}":{"patch":{"tags":["Messages"],"summary":"Editar mensagem","description":"Permite editar o conteúdo de uma mensagem posteriormente enviada","operationId":"editMessage","parameters":[{"name":"mId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"content","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/teste":{"get":{"tags":["User"],"summary":"Testar identificação de ROLE","description":"Teste de acesso geral","operationId":"userAccess","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/search":{"get":{"tags":["User"],"summary":"Buscar usuário","description":"Permite pesquisar por um perfil de um usuário","operationId":"searchUsers","parameters":[{"name":"search","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDetails"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/info/favedTimes":{"get":{"tags":["User"],"summary":"Visualizar número de favoritados","description":"Permite visualizar quantas vezes seus posts foram favoritados ao todo","operationId":"favedTimes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/getAll":{"get":{"tags":["User"],"summary":"Listar perfis","description":"Gera uma lista dos usuários da plataforma","operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDetails"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/getAccount/{id}":{"get":{"tags":["User"],"summary":"Visualizar informações do usuário","description":"Permite visualizar as informações sobre outro usuário","operationId":"loadUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDetails"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/admin/teste":{"get":{"tags":["User"],"summary":"Testar identificação de ROLE (ADMIN)","description":"Teste de acesso para administradores","operationId":"adminAccess","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/account":{"get":{"tags":["User"],"summary":"Visualizar informações da conta","description":"Permite visualizar as informações gerais da sua conta","operationId":"loadAccountInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDetails"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/topPosts/views":{"get":{"tags":["Posts"],"summary":"Listar posts mais vistos","description":"Gera uma lista de posts mais visualizados da plataforma","operationId":"topVieweds","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/topPosts/likes":{"get":{"tags":["Posts"],"summary":"Listar posts mais curtidos","description":"Gera uma lista de posts mais curtidos da plataforma","operationId":"topLikeds","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/topPosts/favs":{"get":{"tags":["Posts"],"summary":"Listar posts mais favoritados","description":"Gera uma lista de posts mais favoritados da plataforma","operationId":"topFaveds","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/search/street":{"get":{"tags":["Posts"],"summary":"Buscar por rua","description":"Permite filtrar a busca de posts por rua desejada","operationId":"searchByStreet","parameters":[{"name":"street","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/search/avenue":{"get":{"tags":["Posts"],"summary":"Buscar por bairro","description":"Permite filtrar a busca de posts por bairro desejado","operationId":"searchByAvenue","parameters":[{"name":"avenue","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/my-posts":{"get":{"tags":["Posts"],"summary":"Listar meus posts","description":"Serve uma lista de posts realizados por você","operationId":"getMyPosts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/my-favs":{"get":{"tags":["Posts"],"summary":"Listar meus favoritos","description":"Serve uma lista de posts favoritados por você","operationId":"getMyFavs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/likedTimes/{id}":{"get":{"tags":["Posts"],"summary":"Visualizar número de likes","description":"Apresenta a contagem de likes de um post","operationId":"likedTimes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/getOne/{id}":{"get":{"tags":["Posts"],"summary":"Careegar detalhes do post","description":"Apresenta as informações de um post por completo","operationId":"getOnePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}}},"security":[{"BearerAuth":[]}]}},"/api/messages/loadChat/{id}":{"get":{"tags":["Messages"],"summary":"Carregar histórico de conversa","description":"Carrega toda a conversa entre você e o outro usuário","operationId":"loadMessages","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/messages/chats":{"get":{"tags":["Messages"],"summary":"Listar contatos","description":"Busca a lista de contatos do usuário","operationId":"getChats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDetails"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/images/{postId}/post/thumb":{"get":{"tags":["Images"],"summary":"Buscar thumb por post","description":"Permite a busca da primeira imagem de um post","operationId":"getFirstImage","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}},"security":[{"BearerAuth":[]}]}},"/api/images/{postId}/post/all":{"get":{"tags":["Images"],"summary":"Buscar todas imagens por post","description":"Permite a busca de todas as imagens de um post","operationId":"getAllImagesByPostId","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ImageResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/images/get/{imageId}":{"get":{"tags":["Images"],"summary":"Carregar imagem","description":"Permite carregar uma imagem","operationId":"getImageById","parameters":[{"name":"imageId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}},"security":[{"BearerAuth":[]}]}},"/api/follow/getFollowings":{"get":{"tags":["Follow"],"summary":"Listar quem você segue","description":"Permite visualizar os usuários que você segue","operationId":"getFollowings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDetails"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/follow/getFollowings/{userId}":{"get":{"tags":["Follow"],"summary":"Listar quem um usuário segue","description":"Permite visualizar os usuários que alguém segue","operationId":"getFollowingsById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDetails"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/follow/getFollowers":{"get":{"tags":["Follow"],"summary":"Listar seus seguidores","description":"Permite visualizar os usuários que te seguem","operationId":"getFollowers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDetails"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/follow/getFollowers/{userId}":{"get":{"tags":["Follow"],"summary":"Listar seguidores de um usuário","description":"Permite visualizar os usuários seguem alguém","operationId":"getFollowersById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDetails"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/feed":{"get":{"tags":["Feed"],"summary":"Listar posts em ordem cronológica","description":"Apresenta a lista de posts no formato de um feed","operationId":"getFeed","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}}}}}},"/api/comments/getComments/{userId}":{"get":{"tags":["Comments"],"summary":"Listar comentários por perfil","description":"Permite buscar a lista de comentários por perfil do usuário","operationId":"getComments","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/comments/getComments/post/{postId}":{"get":{"tags":["Comments"],"summary":"Listar comentários por post","description":"Permite buscar a lista de comentários por post","operationId":"getCommentsByPost","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/admin/info/posts/favedPosts/{userId}":{"get":{"tags":["Admin"],"summary":"Visualizar posts favoritos do usuário","description":"Permite que o administrador visualize os posts favoritos de um user","operationId":"getFavedPostsByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}}}},"security":[{"BearerAuth":[]}]}},"/api/admin/info/number/favedTimes/{postId}":{"get":{"tags":["Admin"],"summary":"Visualizar número de favoritos por post","description":"Permite que o administrador visualize o número de vezes que um post foi favoritado","operationId":"getFavedTimesByPost","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}},"security":[{"BearerAuth":[]}]}},"/api/admin/info/number/favedPosts/{userId}":{"get":{"tags":["Admin"],"summary":"Visualizar número de posts favoritos do usuário","description":"Permite que o administrador visualize o número de posts favoritados de um user","operationId":"getNumberOfFavedPostsByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}},"security":[{"BearerAuth":[]}]}},"/api/admin/info/number/allPosts/favedTimes/{userId}":{"get":{"tags":["Admin"],"summary":"Visualizar número de vezes que os posts daquele usuário foram favoritados","description":"Permite que o administrador visualize o número de vezes que todos os posts de um user foram favoritados na plataforma","operationId":"getFavedTimesByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}},"security":[{"BearerAuth":[]}]}},"/api/user/deleteProfile":{"delete":{"tags":["User"],"summary":"Deletar conta","description":"Permite deletar sua conta da plataforma","operationId":"deleteProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/unlike/{id}":{"delete":{"tags":["Posts"],"summary":"Descurtir post","description":"Remove seu like do post","operationId":"unlikePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/unfav/{id}":{"delete":{"tags":["Posts"],"summary":"Desfavoritar post","description":"Remove o post da sua lista de favoritos","operationId":"unfavPost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/deleteImage/{id}/{imageId}":{"delete":{"tags":["Posts"],"summary":"Remover imagem","description":"Permite remover uma imagem da sua publicação","operationId":"deleteImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"imageId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/posts/delete/{id}":{"delete":{"tags":["Posts"],"summary":"Deletar post","description":"Permite deletar uma publicação realizada por você","operationId":"deletePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/messages/delete/{id}":{"delete":{"tags":["Messages"],"summary":"Deletar mensagem","description":"Permite deletar uma mensagem anteriormente enviada","operationId":"deleteMessage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/follow/unfollow/{followingId}":{"delete":{"tags":["Follow"],"summary":"Deixar de seguir usuário","description":"Permite deixar de seguir o perfil de outro usuário","operationId":"unfollowUser","parameters":[{"name":"followingId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}},"/api/comments/deleteComment/{id}":{"delete":{"tags":["Comments"],"summary":"Deletar comentário","description":"Permite deletar seu comentário","operationId":"deleteComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}}},"components":{"schemas":{"MessageResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"sendedAt":{"type":"string","format":"date-time"},"senderId":{"type":"integer","format":"int64"},"receiverId":{"type":"integer","format":"int64"}}},"CommentRequest":{"type":"object","properties":{"content":{"type":"string","example":"comentário tal"}}},"CommentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int64"},"authorId":{"type":"integer","format":"int64"},"postId":{"type":"integer","format":"int64"}}},"RegisterRequest":{"type":"object","properties":{"cpf":{"type":"string","example":"000.000.000-00"},"phoneNumber":{"type":"string","example":"(00) 90000-0000"},"name":{"type":"string","example":"Fulano de tal"},"email":{"type":"string","example":"fulano@email.com"},"password":{"type":"string","example":"senha123"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","example":"seuemail@email.com"},"password":{"type":"string","example":"senha@123"}}},"PromoteRequest":{"type":"object","properties":{"email":{"type":"string","example":"umemail@email.com"}}},"SetPasswordRequest":{"type":"object","properties":{"password":{"type":"string","example":"senha123"},"newPassword":{"type":"string","example":"Senha@123"}}},"SetInfoRequest":{"type":"object","properties":{"bio":{"type":"string","example":"AAAAAA"},"name":{"type":"string","example":"Um nome criativo"},"phoneNumber":{"type":"string","example":"(11) 91111-0000"}}},"SetPostInfoRequest":{"type":"object","properties":{"description":{"type":"string","example":"Uma casa bonita"},"price":{"type":"number","format":"double","example":1000000},"street":{"type":"string","example":"Rua 15"},"avenue":{"type":"string","example":"Bairro Esperança"},"number":{"type":"string","example":"15"},"type":{"type":"string","example":"venda"}}},"UserDetails":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"imageProfilePath":{"type":"string"},"imageProfileId":{"type":"integer","format":"int64"},"email":{"type":"string"},"name":{"type":"string"},"phoneNumber":{"type":"string"},"bio":{"type":"string"},"role":{"type":"string"},"followers":{"type":"integer","format":"int32"},"followings":{"type":"integer","format":"int32"}}},"PostResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstImageAdresse":{"type":"string"},"allImagesAdresse":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"},"street":{"type":"string"},"avenue":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"number":{"type":"string"},"type":{"type":"string"},"favedTimes":{"type":"integer","format":"int32"},"likedTimes":{"type":"integer","format":"int32"},"views":{"type":"integer","format":"int32"},"reachedTimes":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int64"},"wasFaved":{"type":"boolean"},"wasLiked":{"type":"boolean"},"wasUpdated":{"type":"boolean"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"}}}},"ImageResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"filename":{"type":"string"},"filePath":{"type":"string"},"contentType":{"type":"string"}}},"DeleteProfileRequest":{"type":"object","properties":{"password":{"type":"string"}}}},"securitySchemes":{"BearerAuth":{"type":"http","name":"BearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}